+++ wp-parsely.php
@@ -54,12 +54,22 @@ const PARSELY_FILE    = __FILE__;
 
 require __DIR__ . '/src/class-parsely.php';
 require __DIR__ . '/src/class-scripts.php';
-require __DIR__ . '/src/class-metadata.php';
 require __DIR__ . '/src/class-dashboard-link.php';
 require __DIR__ . '/src/UI/class-admin-bar.php';
 require __DIR__ . '/src/UI/class-metadata-renderer.php';
 require __DIR__ . '/src/Endpoints/class-metadata-endpoint.php';
 require __DIR__ . '/src/Endpoints/class-graphql-metadata.php';
+
+require __DIR__ . '/src/class-metadata.php';
+require __DIR__ . '/src/Meta/class-metadata-builder.php';
+require __DIR__ . '/src/Meta/class-author-archive-builder.php';
+require __DIR__ . '/src/Meta/class-category-builder.php';
+require __DIR__ . '/src/Meta/class-front-page-builder.php';
+require __DIR__ . '/src/Meta/class-page-builder.php';
+require __DIR__ . '/src/Meta/class-page-for-posts-builder.php';
+require __DIR__ . '/src/Meta/class-paginated-front-page-builder.php';
+require __DIR__ . '/src/Meta/class-post-builder.php';
+require __DIR__ . '/src/Meta/class-tag-builder.php';
 
 add_action( 'plugins_loaded', __NAMESPACE__ . '\\parsely_initialize_plugin' );
 /**
